+++ crates/maybenot/CHANGELOG.md
@@ -7,6 +7,8 @@ Manually generated changelog, for now. We follow semantic versioning.
 - Bug fix: actions are now scheduled correctly when a counter is zeroed.
   Previously, a counter being decremented to zero on a state transition
   would result in no action in the absence of further transitions.
+- Bug fix: account for two CounterZero transitions out and back into a state
+  being state changes (that should not prematurely decrement any set limits).
 
 ## 2.0.0 - 2024-09-09